Code cleanup in preparation for fixing GPU detect issue
lib/util.cpp had become a landfill of unrelated stuff
    - move the process-related code to lib/proc_control.cpp
    - move file-related code to lib/filesys.cpp
    - move client mutex code to client/main.cpp

The client has 2 unrelated types of mutual exclusion:
thread level (between the main and throttle threads)
and process level (prevent 2 clients from running on a host).
    - Rename "client_mutex" to "client_thread_mutex".

- parse.h: don't declare boinc_is_finite(); include the .h file
