+++ client/dhrystone.cpp
@@ -47,7 +47,7 @@
 #include <cstring>
 #endif
 
-#include "util.h"
+#include "proc_control.h"
 #include "cpu_benchmark.h"
 #include "dhrystone.h"
