+++ clientgui/MainDocument.cpp
@@ -24,17 +24,14 @@
 
 #include "error_numbers.h"
 #include "str_replace.h"
-#include "util.h"
-#include "BOINCGUIApp.h"
-#include "MainDocument.h"
+#include "proc_control.h"
 
 #ifdef __WXMAC__
 #include "mac_util.h"
 #endif
-#ifdef _WIN32
-#include "proc_control.h"
-#endif
 
+#include "BOINCGUIApp.h"
+#include "MainDocument.h"
 #include "BOINCBaseFrame.h"
 #include "AdvancedFrame.h"
 #include "BOINCClientManager.h"
